Refactor TodayRangeData into separate files
