+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/mystore/StatsRepository.kt
@@ -1,5 +1,6 @@
 package com.woocommerce.android.ui.mystore
 
+import com.woocommerce.android.ui.mystore.stats.TodayRangeData
 import com.woocommerce.commons.extensions.formatToYYYYmmDDhhmmss
 import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.fluxc.model.WCRevenueStatsModel
